split up rfuns_extensions.cpp
