+++ CMakeLists.txt
@@ -9,7 +9,9 @@ set(LOADABLE_EXTENSION_NAME ${TARGET_NAME}_loadable_extension)
 project(${TARGET_NAME})
 include_directories(src/include duckdb/src/include)
 
-set(EXTENSION_SOURCES src/rfuns_extension.cpp)
+set(EXTENSION_SOURCES src/rfuns_extension.cpp 
+                      src/add.cpp
+                      src/eq.cpp)
 
 build_static_extension(${TARGET_NAME} ${EXTENSION_SOURCES})
 build_loadable_extension(${TARGET_NAME} " " ${EXTENSION_SOURCES})
