+++ src/include/rfuns_extension.hpp
@@ -3,6 +3,12 @@
 #include "duckdb.hpp"
 
 namespace duckdb {
+namespace rfuns {
+
+ScalarFunctionSet base_r_add();
+ScalarFunctionSet base_r_eq();
+
+}
 
 class RfunsExtension : public Extension {
 public:
